Finalize UI of game
This is not working exactly as planned. The alert message
is executed before the DOM element of the score is redrawn.
I am not sure how to force the redraw before the alert, but am
hopeful I will learn in upcoming lessons. For now the most
important part of this lesson is using git to merge this branch
back into main. So I will not waste further time here.